# Conflicts:
#	src/FedTree/FL/FLtrainer.cpp
#	src/FedTree/Tree/hist_tree_builder.cpp

+++ src/FedTree/FL/partition.cpp
@@ -37,7 +37,7 @@ Partition::homo_partition(const DataSet &dataset, const int n_parties, const boo
 //Todo add hetero partition according to the labels
 std::map<int, vector<int>>
 Partition::hetero_partition(const DataSet &dataset, const int n_parties, const bool is_horizontal,
-                            vector<double> alpha) {
+                            vector<float> alpha) {
     int n;
     if (is_horizontal)
         n = dataset.n_instances();
@@ -58,10 +58,10 @@ Partition::hetero_partition(const DataSet &dataset, const int n_parties, const b
     std::random_device rd;
     std::mt19937 gen(rd());
     dirichlet_distribution<std::mt19937> d(alpha);
-    vector<double> dirichlet_samples;
-    for (double x : d(gen)) dirichlet_samples.push_back(x);
+    vector<float> dirichlet_samples;
+    for (float x : d(gen)) dirichlet_samples.push_back(x);
     std::transform(dirichlet_samples.begin(), dirichlet_samples.end(), dirichlet_samples.begin(),
-                   [&n](double &c) { return c * n; });
+                   [&n](float &c) { return c * n; });
     std::partial_sum(dirichlet_samples.begin(), dirichlet_samples.end(), dirichlet_samples.begin());
 
     for (auto &x : dirichlet_samples)
@@ -79,7 +79,7 @@ Partition::hetero_partition(const DataSet &dataset, const int n_parties, const b
 }
 
 
-void Partition::hybrid_partition(const DataSet &dataset, const int n_parties, vector<double> alpha,
+void Partition::hybrid_partition(const DataSet &dataset, const int n_parties, vector<float> alpha,
                                  vector<DataSet> &subsets){
     for(int i = 0; i < n_parties; i++){
         //todo: group label
@@ -93,7 +93,7 @@ void Partition::hybrid_partition(const DataSet &dataset, const int n_parties, ve
     int seed = 42;
     std::mt19937 gen(seed);
     dirichlet_distribution<std::mt19937> dir(alpha);
-    vector<double> dir_numbers = dir(gen);
+    vector<float> dir_numbers = dir(gen);
     CHECK_EQ(dir_numbers.size(),n_parties);
     vector<int> n_parts_each_party_accu(n_parties + 1, 0);
     for(int i = 1; i < n_parties; i++){
